Refactor: Moves saveState thunk to epic
